+++ src/commands/command.cpp
@@ -12,7 +12,6 @@ Command::Command(const std::string& name,
     : m_name(name)
     , m_cli(name, description) {
         m_cli.addOption('h', "help", false, "Show help page");
-        m_cli.addOption('d', "debug", false, "Setup debug mode.");
         m_cli.addOption('f', "file", true, "Database file.");
     }
 
@@ -29,17 +28,6 @@ std::string Command::getUsage() const {
 const std::string Command::getDefaultFile() const {
     const std::string username{getenv("USER")};
     return "/home/" + username + "/.cora.sqlite";
-}
-
-void Command::configureLogger(const bool debug) const {
-    if (debug) {
-        static plog::ConsoleAppender<plog::TxtFormatter> consoleAppender;
-        plog::init(plog::debug, &consoleAppender);
-        LOG(plog::debug) << "Debug logging configured";
-    } else {
-        static plog::ConsoleAppender<MessageOnlyFormatter> consoleAppender;
-        plog::init(plog::info, &consoleAppender);
-    }
 }
 
 std::shared_ptr<MediaPlayer> Command::createPlayer() const {