+++ src/commands/import_command.cpp
@@ -1,11 +1,12 @@
 #include "commands/import_command.hpp"
 
+#include <plog/Log.h>
+
 #include "importer/radio_sure_importer.hpp"
 
 #include <memory>
 
-ImportCommand::ImportCommand()
-    : Command("import", "search for radio staions") {
+ImportCommand::ImportCommand() : Command("import", "search for radio staions"), m_importerByName() {
         auto radioSureImporter = std::unique_ptr<RadioSureImporter>(new RadioSureImporter());
         m_importerByName.insert(std::make_pair(radioSureImporter->getName(), std::move(radioSureImporter)));
         m_cli.addOption('i', "input", true, "Input to import stations from. Depends on the type an could be a file or URL.");
@@ -16,7 +17,6 @@ ImportCommand::~ImportCommand() { }
 
 void ImportCommand::execute(const std::vector<std::string>& args) {
     m_cli.parse(args);
-    configureLogger(m_cli.hasValue('d'));
 
     if (m_cli.hasOption('h')) {
         LOG(plog::info) << m_cli.usage();