+++ src/cora.hpp
@@ -8,12 +8,19 @@
 class Cora {
     public:
         Cora();
+
         ~Cora();
+
         void run(int argc, char* argv[]);
+
     private:
         CommandInterpreter m_commandInterpreter;
 
         void addStations();
+
         void runCommand(const std::vector<std::string>& args);
+
+        void configureLogger() const;
+
         void showUsage();
 };